delete package permission.go and move custom_validator.go to components
