--- permission/permission.go
@@ -1,1 +1,0 @@
-package permission
